fix(geom): preserve f64 precision through xform and camera readers
Pixar's UsdGeomXformOp allows xformOp:rotate* to be authored as either float
or double, and UsdGeomCamera declares shutter:open / shutter:close as double.
The readers previously narrowed both to f32 before the math ran, silently
truncating mantissa bits a Pixar-compatible reader would have kept.

xform.rs now keeps rotation angles in f64 end-to-end (value extract,
to_radians, sin_cos, matrix). ReadCamera.shutter_open / shutter_close move to
f64. Other camera attributes stay f32 because their schema type is float.

Also tightens compute_local_to_parent_transform: a !resetXformStack! sentinel
is only valid at index 0 per the UsdGeomXformable spec; a non-leading
occurrence now surfaces an error instead of being silently dropped.

+++ src/schemas/geom/types.rs
@@ -298,8 +298,8 @@ pub struct ReadCamera {
     /// Each entry is a `float4` plane equation `(n.x, n.y, n.z, d)`
     /// in camera-local space.
     pub clipping_planes: Vec<[f32; 4]>,
-    pub shutter_open: f32,
-    pub shutter_close: f32,
+    pub shutter_open: f64,
+    pub shutter_close: f64,
     pub exposure: f32,
     pub exposure_iso: f32,
     pub exposure_time: f32,
